WI #1152 Implement the transpilation of a C# Razor Script to C# instructions (A06.T01)
The transpilation of a C# razor Script to C# is performed using interpolated strings, and Razor meta character substitution in C# code.

WI #1152 Implement the transpilation of a C# Razor Script to C# instructions (A06.T01)

This transpilation is better  as it will use an internal StringBuilder instance to collect Markup text between C# code block.