WI #1301 Optimize variable search under Typedefcontext
Instead of getting a list with all DataDefinition children that match the searched DataDefinition name then search in this list if the searched DataDefinition is present, directly compare the parentTypeDefinition.

The only use case is to resolve DEPENDING ON inside typedef so
performance gain should not be visible in real situation.