chore: Fix build error when cloning for first time
When cloning the repository for the first time and try to build it; many errors occur because of the compiling process requiring unknown generated code insides TypeCobol.csproj.

The proposed solution adds a condition that ensures any generated sources are linked in the build process only if they exists. This fix should not have any impacts on:
- Other development environment
- CI/CD Build process (presumption)

Tests status: OK